# Backend/Syslab/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "pablokauan.tech@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Listar notícias",
                "operationId": "9bfe744cf19dcb58e9ab9677c29ee6c7",
                "responses": {
                    "200": {
                        "description": "Lista de notícias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/News"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Criar notícia",
                "operationId": "1c0fdf6cc53f1508a87121ef9111be5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "body_text",
                                    "author_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "subtitle": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "body_text": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "datastamp": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "author_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notícia criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/News"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Buscar notícia por ID",
                "operationId": "235a202c862ec6888f4129413c1f2bcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notícia encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/News"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notícia não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "News"
                ],
                "summary": "Atualizar notícia",
                "operationId": "11a73338ef61a1ffd2ab3771639f638e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "subtitle": {
                                        "type": "string"
                                    },
                                    "body_text": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "datastamp": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "author_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notícia atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/News"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Remover notícia",
                "operationId": "cd3235328b3ea2080e4d0a47986d4714",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Notícia removida"
                    },
                    "404": {
                        "description": "Notícia não encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "News": {
                "title": "News",
                "description": "Notícia publicada no sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Novo laboratório é inaugurado"
                    },
                    "subtitle": {
                        "type": "string",
                        "example": "Espaço amplia pesquisas na universidade",
                        "nullable": true
                    },
                    "body_text": {
                        "type": "string",
                        "example": "O laboratório foi inaugurado com foco em inovação..."
                    },
                    "image": {
                        "type": "string",
                        "example": "news.jpg",
                        "nullable": true
                    },
                    "datastamp": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-01T10:30:00Z"
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author": {
                        "$ref": "#/components/schemas/Pessoa"
                    }
                },
                "type": "object"
            },
            "Pessoa": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pablo Kauan"
                    },
                    "image": {
                        "type": "string",
                        "example": "perfil.jpg",
                        "nullable": true
                    },
                    "research": {
                        "type": "string",
                        "example": "Inteligência Artificial",
                        "nullable": true
                    },
                    "lattes": {
                        "type": "string",
                        "nullable": true
                    },
                    "orcid": {
                        "type": "string",
                        "nullable": true
                    },
                    "descricao": {
                        "type": "string",
                        "nullable": true
                    },
                    "cargo": {
                        "type": "string",
                        "nullable": true
                    },
                    "curso": {
                        "type": "string",
                        "nullable": true
                    },
                    "projetos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Projeto"
                        }
                    },
                    "news": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/News"
                        }
                    }
                },
                "type": "object"
            },
            "Projeto": {
                "title": "Projeto",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sistema de Monitoramento Energético"
                    },
                    "image": {
                        "type": "string",
                        "example": "projeto.jpg",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Projeto voltado ao monitoramento de consumo energético",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "ativo"
                    },
                    "date_begin": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-01"
                    },
                    "date_end": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31",
                        "nullable": true
                    },
                    "pivot": {
                        "properties": {
                            "integrantes": {
                                "type": "string",
                                "example": "Desenvolvedor Backend",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "pessoas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Pessoa"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}